refactored code and added batch invitation ...
