+++ application.py
@@ -8,10 +8,11 @@ from flask_admin import Admin
 from flask_debugtoolbar import DebugToolbarExtension
 from flask_mail import Mail
 from playhouse import db_url
-##from raven.contrib.flask import Sentry
 
 from config import *  # noqa: F401, F403
 from failover import PgDbWithFailover
+
+##from raven.contrib.flask import Sentry
 
 app = Flask(__name__)
 app.config.from_object(__name__)
