+++ models.py
@@ -10,10 +10,9 @@ from itertools import zip_longest
 from urllib.parse import urlencode
 
 from flask_login import UserMixin
-from peewee import (BooleanField, CharField, CompositeKey, DateTimeField,
-                    DeferredRelation, Field, ForeignKeyField, IntegerField,
-                    Model, OperationalError, SmallIntegerField, TextField,
-                    datetime)
+from peewee import (BooleanField, CharField, CompositeKey, DateTimeField, DeferredRelation, Field,
+                    ForeignKeyField, IntegerField, Model, OperationalError, SmallIntegerField,
+                    TextField, datetime)
 from pycountry import countries
 
 from application import db
